Moving CookieTenantSessionManager to integration-tests module
